refactored my homework8
