+++ apps/first_example/views/users_accounts_view.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render
-from apps.first_example.actions.create_accounts import create_account_for_users_from_generator
+from apps.first_example.actions.create_users import create_generator_of_humans
 
 
 def create_users_accounts(request):
@@ -7,7 +7,7 @@ def create_users_accounts(request):
         request=request,
         template_name="first_example/users.html",
         context={
-            "users": create_account_for_users_from_generator(),
+            "users": create_generator_of_humans(),
         },
     )
 
@@ -16,5 +16,5 @@ def create_users_accounts_amount(request, amount=5):
     return render(
         request=request,
         template_name="first_example/users.html",
-        context={"users": create_account_for_users_from_generator(amount=amount), "amount": amount},
+        context={"users": create_generator_of_humans(amount=amount), "amount": amount},
     )
